requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
PyJWT==2.3.0
python-dotenv==0.19.0
requests==2.26.0
bcrypt==3.2.0
PyMySQL==1.0.2
mysql-connector-python==8.0.26
python-dateutil==2.8.2
geopy==2.2.0
networkx==2.6.3
